+++ src/main/java/com/bookingbustickets/bookingbustickets/controller/response/ResponsePassengerCategoryDto.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,5 +13,5 @@ public class ResponsePassengerCategoryDto {
 
     private String categoryName;
 
-    private BigDecimal discountPercentage;
+    private float discountPercentage;
 }
